+++ services/workflows-service/src/workflow/dtos/workflow-where-unique-input.ts
@@ -1,5 +1,6 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsString } from 'class-validator';
+import { Type } from '@sinclair/typebox';
 
 export class WorkflowDefinitionWhereUniqueInput {
   @ApiProperty({
@@ -9,3 +10,7 @@ export class WorkflowDefinitionWhereUniqueInput {
   @IsString()
   id!: string;
 }
+
+export const WorkflowDefinitionWhereUniqueInputSchema = Type.String({
+  description: "The workflow's id",
+});