+++ services/workflows-service/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @ballerine/workflows-service
 
+## 0.7.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.7
+  - @ballerine/workflow-core@0.6.12
+  - @ballerine/workflow-node-sdk@0.6.12
+
+## 0.7.10
+
+### Patch Changes
+
+- Version bump
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.11
+  - @ballerine/workflow-node-sdk@0.6.11
+
+## 0.7.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/workflow-core@0.6.10
+  - @ballerine/workflow-node-sdk@0.6.10
+
+## 0.7.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.6
+  - @ballerine/workflow-core@0.6.9
+  - @ballerine/workflow-node-sdk@0.6.9
+
 ## 0.7.7
 
 ### Patch Changes